pyproject.toml
[project]
name = "hasty-feedback-system"
version = "2.0.0"
description = "A feedback system with SQLite and Tortoise ORM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "tortoise-orm>=0.20.0",
    "aiosqlite>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]
